Move startup into config
